refactor(1682 #1): rename safety/ → limits/ (resolve safe/-vs-safety/ collision)
#1 structure cleanup (#1682): the `safe/` (FP-0042 allowlist) vs `safety/`
(runtime limits) name-collision is indistinguishable to a reader. Rename the
runtime-limits package `safety/` → `limits/` (lead's pick). `safe/` is UNTOUCHED
(allowlist-coupled, owner-gated).

Approach B (rewrite + delete, not a shim — lead-confirmed): the surface is tiny +
internal-only (11 deep `from reyn.safety.limit_handler import` callers; NO
skill/op-catalog/config surface, verified per the rename-drift pin), AND a
lingering safety/ shim would defeat the goal (eliminate the colliding name). So:
- git-mv safety → limits (limit_handler.py + __init__.py).
- rewrite the 11 deep module imports reyn.safety.limit_handler →
  reyn.limits.limit_handler (chat/{planner,session,router_loop,services/a2a_handler},
  kernel/{phase_executor,run_orchestrator}, + 5 tests). Same symbols, new path.
- limits/__init__ internal import repointed; safety/ DELETED (no shim).
- CONFIG-key namespace `safety.on_limit` / `safety.loop` (reyn.yaml) is UNCHANGED —
  that's the config, not the module (the __init__ docstring's safety.on_limit
  prose correctly stays).
- Historical proposal docs (0005-safety-as-checkpoint) left as-is (deep-dive
  history, no churn — lead-confirmed).

Verified: reyn.limits + public surface (LimitDecision/handle_limit_exceeded/
reset_run_extensions) + all 11 importers import OK; 19 limit-handler tests pass;
ruff src tests clean (1 I001 auto-fixed). behavior-preserving.

Refs #1682 (#1).
