+++ tests/test_safety_phase2_wiring.py
@@ -26,7 +26,7 @@ from reyn.budget.budget import BudgetTracker, CostConfig
 from reyn.chat.session import ChatSession, RouterCapExceeded
 from reyn.config import LoopConfig, OnLimitConfig, SafetyConfig
 from reyn.kernel.runtime import LoopLimitExceededError, OSRuntime
-from reyn.safety.limit_handler import reset_run_extensions
+from reyn.limits.limit_handler import reset_run_extensions
 from reyn.schemas.models import Phase, Skill, SkillGraph
